+++ app/src/main/java/com/jdemkiewicz/opentripplanner/ApiModel/Leg.java
@@ -13,10 +13,10 @@ public class Leg {
 
     @SerializedName("startTime")
     @Expose
-    private Integer startTime;
+    private Long startTime;
     @SerializedName("endTime")
     @Expose
-    private Integer endTime;
+    private Long endTime;
     @SerializedName("departureDelay")
     @Expose
     private Integer departureDelay;
@@ -93,19 +93,19 @@ public class Leg {
     @Expose
     private String routeShortName;
 
-    public Integer getStartTime() {
+    public Long getStartTime() {
         return startTime;
     }
 
-    public void setStartTime(Integer startTime) {
+    public void setStartTime(Long startTime) {
         this.startTime = startTime;
     }
 
-    public Integer getEndTime() {
+    public Long getEndTime() {
         return endTime;
     }
 
-    public void setEndTime(Integer endTime) {
+    public void setEndTime(Long endTime) {
         this.endTime = endTime;
     }
